refactor(jstzd): refactor tests
